Move avatar form to its own component
